いくつかのStageLightPropertyに対してPropertyDrawerを実装
