+++ jp.iridescenet.stagelightmaneuver/Editor/PropertyDrawer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9a34590ee7a8484abdc41a01e69acc2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
